+++ src/Encamina.Enmarcha.AspNet.Mvc/Bindings/CustomDateTimeModelBinderProvider.cs
@@ -8,7 +8,7 @@ namespace Encamina.Enmarcha.AspNet.Mvc.Bindings;
 /// </summary>
 public class CustomDateTimeModelBinderProvider : IModelBinderProvider
 {
-    private readonly string customDateTimeFormat;
+    private readonly string? customDateTimeFormat;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="CustomDateTimeModelBinderProvider"/> class.
@@ -21,16 +21,16 @@ public class CustomDateTimeModelBinderProvider : IModelBinderProvider
     /// Initializes a new instance of the <see cref="CustomDateTimeModelBinderProvider"/> class with specific custom date time format.
     /// </summary>
     /// <param name="customDateTimeFormat">The custom <see cref="DateTime"/> format to use when binding models.</param>
-    public CustomDateTimeModelBinderProvider(string customDateTimeFormat)
+    public CustomDateTimeModelBinderProvider(string? customDateTimeFormat)
     {
         this.customDateTimeFormat = customDateTimeFormat;
     }
 
     /// <inheritdoc/>
-    public virtual IModelBinder GetBinder(ModelBinderProviderContext context)
+    public virtual IModelBinder? GetBinder(ModelBinderProviderContext context)
     {
         return CustomDateTimeModelBinder.SupportedDateTimeTypes.Contains(context.Metadata.ModelType)
             ? new CustomDateTimeModelBinder(customDateTimeFormat)
-            : (IModelBinder)null;
+            : (IModelBinder?)null;
     }
 }